support resource level kubernetes configuration.
This patch introduces support for providing kubernetes configuration at the
resource level as below:

```hcl
resource "kustomization_resource" "from_overlay" {
  for_each = data.kustomization_overlay.test.ids

  manifest        = data.kustomization_overlay.test.manifests[each.value]
  kubeconfig_path = "~/.kube/config"
}
```

or:

```hcl
resource "kustomization_resource" "from_overlay" {
  for_each = data.kustomization_overlay.test.ids

  manifest        = data.kustomization_overlay.test.manifests[each.value]
  kubeconfig_raw  = <<EOF
replace_with_kube_config_content
EOF
}
```

## Why is this needed

### case 1

When the provider is initialized with a kubeconfig dynamically e.g from
another module as below:

```hcl
module "kubernetes_cluster" {
  source = "some_module_name"
  # some variables here, but the module outputs a `kubeconfig`
}

provider "kustomization" {
  kubeconfig_raw = module.kubernetes_cluster.kubeconfig
}
```

in some situations terraform will initialize the provider before the modules
output is available and set `kubeconfig_raw` as an empty string. See
https://github.com/hashicorp/terraform/issues/2430#:~:text=commented-,on%20Feb%2028,-%E2%80%A2
for more details on this. When the provider receives an empty string as its
configuration the kubernetes client-go falls back to its default client
initialization and fails with:

```
Error: Get "http://localhost/api?timeout=32s": dial tcp [::1]:80: connect: connection refused
```

With this patch, the resource dependency is explicit, and terraform will not
proceed with calling the client until the module providing the dynamic
kubeconfig is completely resolved.

### case 2

When this provider and its resources are embedded in a module, it is not
possible to use the module in a `for_each` loop to target different clusters.
Take the previous example in case one with variables `kubeconfig` and
`cluster_name`, if one attempts to use it in a for_each loop as below:

```hcl

locals = {
  clusters = {
    "one": "kube_config_for_cluster_one",
    "two": "kubeconfig_for_cluster_2",
  }
}

module "wrapping_cluster_and_kustomization" {
  for_each = local.clusters

  kubeconfig = each.value
  cluster_name = each.key
}
```

The above fails with the below error:

```
The module at module.wrapping_cluster_and_kustomization is a legacy module which contains its own local provider configurations, and so calls to it may not use the count, for_each, or depends_on arguments.
```

Moving the kubeconfig into the resource definition means we can have a root
provider with `kubeconfig_raw` configured to an empty string while the module
sets the `kubeconfig` passed into it to the resources and can do without the
provider configuration.

## drawbacks

Each resource definition will trigger at least 2 API calls to the apiserver,
one for Read and the other for Exists, unlike provider level configuration
that reuses the same TCP connection.